+++ src/main/java/ru/vsu/entity/mappers/AttributeMapper.java
@@ -9,9 +9,9 @@ import java.sql.SQLException;
 public class AttributeMapper implements RowMapper<AttributeEntity> {
     @Override
     public AttributeEntity mapRow(ResultSet resultSet, int i) throws SQLException {
-        return new AttributeEntity(resultSet.getBigDecimal("id").toBigInteger(),
+        return new AttributeEntity(resultSet.getLong("id"),
                 resultSet.getString("name"),
-                resultSet.getBigDecimal("object_type_id").toBigInteger(),
+                resultSet.getLong("object_type_id"),
                 resultSet.getString("value_type"),
                 resultSet.getBoolean("require"));
