+++ src/main/java/ru/vsu/entity/mappers/ObjectTypeMapper.java
@@ -3,14 +3,13 @@ package ru.vsu.entity.mappers;
 import org.springframework.jdbc.core.RowMapper;
 import ru.vsu.entity.ObjectTypeEntity;
 
-
 import java.sql.ResultSet;
 import java.sql.SQLException;
 
 public class ObjectTypeMapper implements RowMapper<ObjectTypeEntity> {
     @Override
     public ObjectTypeEntity mapRow(ResultSet resultSet, int i) throws SQLException {
-        return new ObjectTypeEntity(resultSet.getBigDecimal("id").toBigInteger(),
+        return new ObjectTypeEntity(resultSet.getLong("id"),
                 resultSet.getString("name"));
     }
 }
